Refactored validation in different file and other modifications
